[CP-2339] useWatchDeviceDataEffect-* moved to overview
